pipeline/requirements.txt
google-cloud-bigquery==3.25.*
google-cloud-storage==2.18.*
google-cloud-aiplatform==1.72.*
functions-framework==3.*
jinja2==3.1.*
pyyaml==6.*
pytest==8.*
pytest-mock==3.*
